requirements.txt
streamlit==1.27.0
Pillow==9.4.0
opencv==4.7.0
numpy==1.22.4
scikit-image==0.19.3